Collapse summary and dashboard wizard/runner modules into their command modules
summary_wizard.py and dashboard_wizard.py were near-identical 28-line
modules that just deferred to pick_months + load_snapshot. dashboard_runner.py
was a 22-line module with 3 lines of orchestration.

These are now private functions within summary.py and dashboard.py, keeping
the injectable-callback test seam. No behaviour changes.
