--- git_dev_metrics/cli/dashboard_wizard.py
@@ -1,28 +1,0 @@
-from collections.abc import Callable
-from pathlib import Path
-
-import typer
-
-from ..metrics.loader import load_snapshot_for_months
-from ._wizard import _prompt_months, pick_months
-from .dashboard_runner import write_and_open_dashboard
-
-YearMonth = tuple[int, int]
-
-
-def dashboard_wizard(
-    db_path: Path | None = None,
-    *,
-    ask_months: Callable[[list[YearMonth]], list[YearMonth]] = _prompt_months,
-) -> None:
-    """Pick months from cache, render HTML dashboard, open in browser."""
-    selected = pick_months(db_path, ask_months)
-    snapshot = load_snapshot_for_months(selected, db_path)
-    if snapshot is None:
-        typer.secho("No PRs in selected months.", fg=typer.colors.RED, err=True)
-        raise typer.Exit(code=1)
-    first, last = selected[0], selected[-1]
-    period_slug = f"{first[0]:04d}-{first[1]:02d}-to-{last[0]:04d}-{last[1]:02d}"
-    since = snapshot.period.since.strftime("%Y-%m-%d")
-    until = snapshot.period.until.strftime("%Y-%m-%d")
-    write_and_open_dashboard(snapshot, period_slug, f"{since} to {until}", output=None)
